Move navigation to an effect
