+++ src/app/modules/layout/nav-bar/nav-bar.component.ts
@@ -1,13 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { NavigationEnd, Router } from '@angular/router';
-import { Actions, ofType } from '@ngrx/effects';
+import { Actions } from '@ngrx/effects';
 import { Store } from '@ngrx/store';
-import { combineLatest, filter, switchMap, tap, withLatestFrom } from 'rxjs';
-import { APIUserResponseDTO } from 'src/app/api/v2';
+import { combineLatest, filter, tap } from 'rxjs';
 import { BaseComponent } from 'src/app/shared/base/base.component';
-import { StartPreferenceChoice } from 'src/app/shared/models/organization/organization-user/start-preference.model';
-import { UIRootConfig } from 'src/app/shared/models/ui-config/ui-root-config.model';
 import { filterNullish } from 'src/app/shared/pipes/filter-nullish';
 import { OrganizationActions } from 'src/app/store/organization/actions';
 import { selectUIRootConfig } from 'src/app/store/organization/selectors';
@@ -40,44 +37,7 @@ export class NavBarComponent extends BaseComponent implements OnInit {
   }
 
   ngOnInit(): void {
-    this.setupUseUserDefaultStartPage();
     this.setupGetUIRootConfigOnNavigation();
-  }
-
-  private setupUseUserDefaultStartPage() {
-    this.subscriptions.add(
-      this.actions$
-        .pipe(
-          ofType(UserActions.resetOnOrganizationUpdate),
-          switchMap(() =>
-            this.uiRootConfig$.pipe(
-              filter((config) => !!config),
-              withLatestFrom(this.user$)
-            )
-          )
-        )
-        .subscribe(([uiRootConfig, user]) => {
-          const userDefaultStartPage = user?.defaultStartPage;
-          if (this.shouldGoToUserDefaultStartPage(userDefaultStartPage, uiRootConfig)) {
-            this.navigateToUserDefaultStartPage(userDefaultStartPage!);
-          }
-        })
-    );
-  }
-
-  private shouldGoToUserDefaultStartPage(
-    userDefaultStartPage: StartPreferenceChoice | undefined,
-    uiRootConfig: UIRootConfig
-  ): boolean {
-    return (
-      this.isOnStartPage() &&
-      userDefaultStartPage !== undefined &&
-      !this.userDefaultStartPageDisabledInOrganization(userDefaultStartPage, uiRootConfig)
-    );
-  }
-
-  private isOnStartPage(): boolean {
-    return this.router.url.replace('/', '') === AppPath.root;
   }
 
   private setupGetUIRootConfigOnNavigation() {
@@ -89,49 +49,6 @@ export class NavBarComponent extends BaseComponent implements OnInit {
         )
         .subscribe()
     );
-  }
-
-  private userDefaultStartPageDisabledInOrganization(
-    userDefaultStartPage: StartPreferenceChoice,
-    uiRootConfig: UIRootConfig
-  ): boolean {
-    const startPageValue = userDefaultStartPage.value;
-    switch (startPageValue) {
-      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.ItSystemCatalog:
-      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.ItSystemUsage:
-        return !uiRootConfig.showItSystemModule;
-      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.ItContract:
-        return !uiRootConfig.showItContractModule;
-      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.DataProcessing:
-        return !uiRootConfig.showDataProcessing;
-      default:
-        return false;
-    }
-  }
-
-  private navigateToUserDefaultStartPage(userDefaultStartPage: StartPreferenceChoice) {
-    const path = this.getUserDefaultStartPagePath(userDefaultStartPage);
-    this.router.navigate([path]);
-  }
-
-  private getUserDefaultStartPagePath(userDefaultStartPage: StartPreferenceChoice): string {
-    const startPageValue = userDefaultStartPage.value;
-    switch (startPageValue) {
-      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.StartSite:
-        return AppPath.root;
-      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.Organization:
-        return `${AppPath.organization}/${AppPath.structure}`;
-      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.ItSystemCatalog:
-        return `${AppPath.itSystems}/${AppPath.itSystemCatalog}`;
-      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.ItSystemUsage:
-        return `${AppPath.itSystems}/${AppPath.itSystemUsages}`;
-      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.ItContract:
-        return AppPath.itContracts;
-      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.DataProcessing:
-        return AppPath.dataProcessing;
-      default:
-        throw new Error(`Unknown start page: ${startPageValue}`);
-    }
   }
 
   public showOrganizationDialog() {
