+++ src/app/store/user-store/effects.ts
@@ -4,9 +4,9 @@ import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { concatLatestFrom } from '@ngrx/operators';
 import { Store } from '@ngrx/store';
 import { CookieService } from 'ngx-cookie';
-import { catchError, combineLatestWith, map, mergeMap, of, switchMap, tap } from 'rxjs';
+import { catchError, combineLatestWith, filter, map, mergeMap, of, switchMap, tap, withLatestFrom } from 'rxjs';
 import { APIUserDTOApiReturnDTO, APIV1AuthorizeINTERNALService } from 'src/app/api/v1';
-import { APIOrganizationGridPermissionsResponseDTO, APIV2PasswordResetInternalINTERNALService } from 'src/app/api/v2';
+import { APIOrganizationGridPermissionsResponseDTO, APIUserResponseDTO, APIV2PasswordResetInternalINTERNALService } from 'src/app/api/v2';
 import { APIV2OrganizationGridInternalINTERNALService } from 'src/app/api/v2/api/v2OrganizationGridInternalINTERNAL.service';
 import { APIV2OrganizationsInternalINTERNALService } from 'src/app/api/v2/api/v2OrganizationsInternalINTERNAL.service';
 import { AppPath } from 'src/app/shared/enums/app-path';
@@ -14,7 +14,10 @@ import { adaptUser } from 'src/app/shared/models/user.model';
 import { filterNullish } from 'src/app/shared/pipes/filter-nullish';
 import { resetOrganizationStateAction, resetStateAction } from '../meta/actions';
 import { UserActions } from './actions';
-import { selectOrganizationUuid } from './selectors';
+import { selectOrganizationUuid, selectUser } from './selectors';
+import { selectUIRootConfig } from '../organization/selectors';
+import { StartPreferenceChoice } from 'src/app/shared/models/organization/organization-user/start-preference.model';
+import { UIRootConfig } from 'src/app/shared/models/ui-config/ui-root-config.model';
 
 @Injectable()
 export class UserEffects {
@@ -96,6 +99,84 @@ export class UserEffects {
       map(() => resetOrganizationStateAction())
     );
   });
+
+  useUserDefaultStartPageOnLogin$ = createEffect(() => {
+    return this.actions$.pipe(
+      ofType(UserActions.resetOnOrganizationUpdate),
+      switchMap(() =>
+        this.store.select(selectUIRootConfig).pipe(
+          filter((config) => !!config),
+          withLatestFrom(this.store.select(selectUser))
+        )
+      ),
+      tap(([uiRootConfig, user]) => {
+        const userDefaultStartPage = user?.defaultStartPage;
+        if (this.shouldGoToUserDefaultStartPage(userDefaultStartPage, uiRootConfig)) {
+          this.navigateToUserDefaultStartPage(userDefaultStartPage!);
+        }
+      })
+    );
+  }, { dispatch: false });
+
+
+
+  private shouldGoToUserDefaultStartPage(
+    userDefaultStartPage: StartPreferenceChoice | undefined,
+    uiRootConfig: UIRootConfig
+  ): boolean {
+    return (
+      this.isOnStartPage() &&
+      userDefaultStartPage !== undefined &&
+      !this.userDefaultStartPageDisabledInOrganization(userDefaultStartPage, uiRootConfig)
+    );
+  }
+
+  private isOnStartPage(): boolean {
+    return this.router.url.replace('/', '') === AppPath.root;
+  }
+
+  private userDefaultStartPageDisabledInOrganization(
+    userDefaultStartPage: StartPreferenceChoice,
+    uiRootConfig: UIRootConfig
+  ): boolean {
+    const startPageValue = userDefaultStartPage.value;
+    switch (startPageValue) {
+      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.ItSystemCatalog:
+      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.ItSystemUsage:
+        return !uiRootConfig.showItSystemModule;
+      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.ItContract:
+        return !uiRootConfig.showItContractModule;
+      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.DataProcessing:
+        return !uiRootConfig.showDataProcessing;
+      default:
+        return false;
+    }
+  }
+
+  private navigateToUserDefaultStartPage(userDefaultStartPage: StartPreferenceChoice) {
+    const path = this.getUserDefaultStartPagePath(userDefaultStartPage);
+    this.router.navigate([path]);
+  }
+
+  private getUserDefaultStartPagePath(userDefaultStartPage: StartPreferenceChoice): string {
+    const startPageValue = userDefaultStartPage.value;
+    switch (startPageValue) {
+      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.StartSite:
+        return AppPath.root;
+      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.Organization:
+        return `${AppPath.organization}/${AppPath.structure}`;
+      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.ItSystemCatalog:
+        return `${AppPath.itSystems}/${AppPath.itSystemCatalog}`;
+      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.ItSystemUsage:
+        return `${AppPath.itSystems}/${AppPath.itSystemUsages}`;
+      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.ItContract:
+        return AppPath.itContracts;
+      case APIUserResponseDTO.DefaultUserStartPreferenceEnum.DataProcessing:
+        return AppPath.dataProcessing;
+      default:
+        throw new Error(`Unknown start page: ${startPageValue}`);
+    }
+  }
 
   goToRootOnAuthenticateFailed$ = createEffect(
     () => {
